requirements.txt
numpy>=1.21.0,<2.0.0
paho-mqtt>=1.6.0
PyQt5>=5.15.0,<5.16.0
opencv-python-headless>=4.5.0,<4.9.0
scipy>=1.7.0
matplotlib>=3.5.0
pyyaml>=6.0
paramiko>=2.8.0
icecream>=2.1.0    # just for debugging